[S11P21D109-294] - 입찰 응답과 경매 로그 반환 타입이 일치 되도록 수정

See merge request s11-blockchain-nft-sub1/S11P21D109!144

+++ backend/spring-server/src/main/java/com/semonemo/spring_server/domain/auction/dto/response/AuctionEndDTO.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 @Builder
 public class AuctionEndDTO {
     private long auctionId;
-    private int finalPrice;
+    private long finalPrice;
     private long winner;
     private String endTime;
 }
